+++ server/src/main/java/org/allaymc/server/AllayServer.java
@@ -26,6 +26,7 @@ import org.allaymc.server.eventbus.AllayEventBus;
 import org.allaymc.server.network.AllayNetworkInterface;
 import org.allaymc.server.player.AllayEmptyPlayerStorage;
 import org.allaymc.server.player.AllayNBTFilePlayerStorage;
+import org.allaymc.server.player.AllayOfflinePlayerManager;
 import org.allaymc.server.player.AllayPlayerManager;
 import org.allaymc.server.plugin.AllayPluginManager;
 import org.allaymc.server.scheduler.AllayScheduler;
@@ -89,7 +90,13 @@ public final class AllayServer implements Server {
 
     private AllayServer() {
         this.state = new AtomicReference<>(ServerState.STARTING);
-        this.playerManager = new AllayPlayerManager(SETTINGS.storageSettings().savePlayerData() ? new AllayNBTFilePlayerStorage(Path.of("players")) : AllayEmptyPlayerStorage.INSTANCE, new AllayNetworkInterface(this));
+        var playersPath = Path.of("players");
+        var playerStorage = SETTINGS.storageSettings().savePlayerData() ? new AllayNBTFilePlayerStorage(playersPath) : AllayEmptyPlayerStorage.INSTANCE;
+        this.playerManager = new AllayPlayerManager(
+                playerStorage,
+                new AllayOfflinePlayerManager(playersPath.resolve("index"), playerStorage),
+                new AllayNetworkInterface(this)
+        );
         this.worldPool = new AllayWorldPool();
         this.computeThreadPool = createComputeThreadPool();
         this.virtualThreadPool = Executors.newVirtualThreadPerTaskExecutor();